Adds hash compatibility
Adds hashes to constraints file to improve integrity and to test that it works.
To test this `skip_install` is set for the canary environment.

Changes the variable that is being patched because patching `install_command` would
fail when the package under test is not listed among the constraints, which it never
is.

Changes order of patch and export to prevent patched dependencies to show up in
exports.