Makes canary less aggressive
... to make tests fail in the commands step

Makes it possible to invoke any other testenv in the coalmine testenv to facilitate
testing various versions.

Removes explicit default arguments to be consistent. It is no more critical for canary
that package under test is installed than it is for any other testenv.

Specifies that canary should always be recreated in the testenv itself to make it
easier to use outside of the coalmine testenv. Keeps the recreate flag in the coalmine
because the coalmine is only useful if package under test is present when the testenv
is created.

Removes canary from default environments because it only needs to be invoked once and
it is safer to do in the coalmine.