Removes misleading links from sample project
Misleading in the sense that they point to locations that as if they could refer to
this project but do no.

Links that point to generic locations, such as the cookiecutter template, are not
considered misleading. Neither is links pointing to locations namespaced under my
username as I can make sure these either do not exist or are correct.

Does not remove broken links because the sample should be as intact as possible.