Replaces canary package with function to generate ganary packages
... for fun and to make it easier publshing new canaries

Multiple canaries will allow install_requires and extras_require to be tested by the
same sample package by depending on different canaries.

Changes the way the canary fails; now it fails on installation so that even if one
forgets to invoke the entry point and it only signals success when the entry point is
executed.