Implements workaround for absolute paths in constraints
The previous workaround with replacing deps only works if the package is a direct
dependency of the environment or the package under test as indirect dependencies
are not listed explicitly in the tox invocation.

Upgrades pytest because the previous version did not have type hints and that made
mypy complain.