Enforce constraints on virtualenv seed packages
... to facilitate using constraints files that are not supported by the most recent
versions of pip.

Another workaround for this is to keep the version of virtualenv installed sufficiently
low that its bundled version of pip is compatible. This should work in CI most of the
time but may fail on developer machines if virtualenv decides to upgrade the pip
version it uses, which it does regularly unless explicitly instructed not to. It can
also be done manually.

Beware that it is easy to corrupt the virtualenv app_data so that it will silently
install the wrong version of seeds.

Upgrade virtualenv because otherwise some of the CLI arguments were not accepted.

Document pip incompatibility because newer versions do not work with hashes; they
complain that there are no hashes for the package being installed yet prints the
same version and hash as is specified in the constraints file.