Rewrite for PIP_CONSTRAINT
PIP_CONSTRAINT is better in multiple ways
* applies to build system requirements
* easier to use with .project

Drop support for hashes in constraints files.
The support in pip has been lacking for a while and I have stopped
using them. Additionally they afford relatively little additional
protection since one often has a cache and/or a private index.
This simplifies the code since it is no longer necessary to install
dependencies ahead of the package under test (PUT).

Drop support for -c.
This avoids adding complexity for supporting multiple paths.
Partially overriding one file with another (as this project did with
the *-lowest environments) does not seem to work in recent versions of
pip anyway.

Allow --requirementsfile to be used while opting out of the other hooks
in case someone wants to extract deps but don't want to hooks that
mess with their tox environments.
Would be nice to factor this out to a command to completely decouple it
from hooks.

Updates package versions. Primarily because dependabot complains about
pip and upgrading pip causes hell to break loose. Blacklist one pip
version that is incompatible with pip-tools.

Remove git filters since these were targeting a niche monorepo use of
mine. I have since starting building distributions ahead of running tox
and creating a temporary constraints file that point to them (it just
appends links to them to the normal constraints file, hacky but simple
and works).